update, add some feature

+++ src/console/Console.php
@@ -2,9 +2,10 @@
 
 namespace swoft\console;
 
+use inhere\console\io\Input;
+use inhere\console\utils\Show;
+
 /**
- *
- *
  * @uses      Console
  * @version   2017年04月25日
  * @author    stelin <phpcrazy@126.com>
@@ -17,33 +18,42 @@ trait Console
     private $status;
     private $command;
 
-    public function parseCommand($args)
+    public function parseCommand()
     {
-        $this->loadSwoftIni();
-        if (!isset($args[1])) {
-            exit("Usage: swoft {start|stop|reload|restart}\n");
+        $input = new Input;
+        $command = $input->getCommand();
+
+        if (!$command || $command === 'help' || $input->getSameOpt(['h', 'help'])) {
+            $this->showHelp($input);
         }
 
-        $command = trim($args[1]);
-        $this->status['startFile'] = $args[0];
-        $methodName = strtolower($command);
-        if(method_exists($this, $methodName) == false){
-            exit("Usage: swoft invalid option: '".$command."'\n");
+        $this->loadSwoftIni();
+
+        $this->status['startFile'] = $input->getScript();
+
+        if (!in_array($command, self::ALLOW_COMMANDS, true)) {
+            Show::error("The command: $command is not exists.");
+            $this->showHelp($input);
         }
 
         $this->command = $command;
-
-        $this->checkStatus();
-        $this->$methodName();
+        $this->$command();
     }
 
-    private function stop()
+    /**
+     * stop the swoole application server
+     */
+    public function stop()
     {
+        if (!$this->isRunning()) {
+            echo "The server is not running! cannot stop\n";
+            exit(0);
+        }
 
         $pidFile = $this->server['pfile'];
         $startFile = $this->status['startFile'];
         @unlink($pidFile);
-        echo("swoft ".$startFile." is stoping ... \n");
+        echo("swoft $startFile is stopping ... \n");
 
         $this->server['masterPid'] && posix_kill($this->server['masterPid'], SIGTERM);
 
@@ -52,33 +62,51 @@ trait Console
 
         while (1) {
             $masterIslive = $this->server['masterPid'] && posix_kill($this->server['masterPid'], SIGTERM);
+
             if ($masterIslive) {
                 if (time() - $startTime >= $timeout) {
-                    echo("swoft ".$startFile." stop fail \n");
+                    echo("swoft " . $startFile . " stop fail \n");
                     exit;
                 }
                 usleep(10000);
                 continue;
             }
-            echo("swoft ".$startFile." stop success \n");
+
+            echo("swoft $startFile stop success \n");
             break;
         }
     }
 
-    private function reload()
+    /**
+     * reload the swoole application server
+     */
+    public function reload()
     {
+        if (!$this->isRunning()) {
+            echo "The server is not running! cannot reload\n";
+            exit(0);
+        }
+
         $startFile = $this->status['startFile'];
-        echo("swoft ".$startFile." is reloading \n");
+
+        echo "swoft $startFile  is reloading \n";
+
         posix_kill($this->server['managerPid'], SIGUSR1);
-        echo("swoft ".$startFile." reload success \n");
+
+        echo "swoft $startFile reload success \n";
     }
 
-    private function restart()
+    /**
+     * restart the swoole application server
+     */
+    public function restart()
     {
-        $this->stop();
+        if ($this->isRunning()) {
+            $this->stop();
+        }
+
         $this->start();
     }
-
 
     private function checkStatus()
     {
@@ -93,31 +121,52 @@ trait Console
             $masterIslive = $this->server['masterPid'] && @posix_kill($this->server['managerPid'], 0);
         }
 
-        if($masterIslive && $this->command == 'start'){
+        if ($masterIslive && $this->command == 'start') {
 //            echo("ysf ".$this->status['startFile']." is already running \n");
 //            exit;
         }
 
-        if($masterIslive == false && $this->command != "start"){
-            echo("ysf ".$this->status['startFile']." is not running \n");
+        if ($masterIslive == false && $this->command != "start") {
+            echo("ysf " . $this->status['startFile'] . " is not running \n");
             exit;
         }
+    }
+
+    /**
+     * check Status
+     * @return bool
+     */
+    protected function isRunning()
+    {
+        $masterIsLive = false;
+        $pFile = $this->server['pfile'];
+
+        if (file_exists($pFile)) {
+            $pidFile = file_get_contents($pFile);
+            $pids = explode(',', $pidFile);
+
+            $this->server['masterPid'] = $pids[0];
+            $this->server['managerPid'] = $pids[1];
+            $masterIsLive = $this->server['masterPid'] && @posix_kill($this->server['managerPid'], 0);
+        }
+
+        return $masterIsLive;
     }
 
     private function loadSwoftIni()
     {
         $setings = parse_ini_file($this->settingPath, true);
-        if(!isset($setings['tcp'])){
+        if (!isset($setings['tcp'])) {
 
         }
-        if(!isset($setings['http'])){
+        if (!isset($setings['http'])) {
 
         }
-        if(!isset($setings['server'])){
+        if (!isset($setings['server'])) {
 
         }
 
-        if(!isset($setings['setting'])){
+        if (!isset($setings['setting'])) {
 
         }
 
@@ -125,5 +174,26 @@ trait Console
         $this->http = $setings['http'];
         $this->server = $setings['server'];
         $this->setting = $setings['setting'];
+    }
+
+    protected function showHelp(Input $input)
+    {
+        $script = $input->getScriptName();
+
+        Show::helpPanel([
+            Show::HELP_DES => 'the application server powered by swoole',
+            Show::HELP_USAGE => "$script <cyan>{start|stop|reload|restart}</cyan> [--opt ...]",
+            Show::HELP_COMMANDS => [
+                'start' => 'start the swoole application server',
+                'restart' => 'restart the swoole application server',
+                'reload' => 'reload the swoole application server',
+                'stop' => 'stop the swoole application server',
+                'help' => 'display the help information',
+            ],
+            Show::HELP_OPTIONS => [
+                '-h,--help' => 'display the help information',
+                '--only-task' => 'only reload task worker when exec reload command'
+            ]
+        ]);
     }
 }
